Added general layout for OBJ 3 & 3A
still not crazy comfortable with python yet so I put a general outline of some kind for objectives 3 & 3a. Doubt this is the ideal way to accomplish this but the general idea should work. (there are some parts that I "commented" with // that need to be looked at and implemented